Add some more test coverage for service injection into various types.
