+++ subprojects/core/src/integTest/groovy/org/gradle/api/ObjectExtensionInstantiationIntegrationTest.groovy
@@ -16,17 +16,9 @@
 
 package org.gradle.api
 
-import org.gradle.api.file.FileSystemOperations
-import org.gradle.api.file.ProjectLayout
-import org.gradle.api.model.ObjectFactory
-import org.gradle.api.provider.ProviderFactory
+
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.internal.execution.WorkExecutor
-import org.gradle.process.ExecOperations
 import spock.lang.Unroll
-
-import javax.inject.Inject
-
 
 class ObjectExtensionInstantiationIntegrationTest extends AbstractIntegrationSpec {
     // Document current behaviour
@@ -400,185 +392,6 @@ class ObjectExtensionInstantiationIntegrationTest extends AbstractIntegrationSpe
 
         expect:
         succeeds()
-    }
-
-    // Document current behaviour
-    def "can inject service and configuration as constructor args when constructor not annotated with @Inject"() {
-        buildFile << """
-            class Thing {
-                Thing(String a, ObjectFactory objects, int b) {
-                    assert a == "a"
-                    assert b == 12
-                    assert objects != null
-                }
-            }
-
-            extensions.create("thing", Thing, "a", 12)
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "can inject service using getter"() {
-        buildFile << """
-            import ${Inject.name}
-
-            class Thing {
-                Thing(String a) {
-                }
-
-                @Inject
-                ObjectFactory getObjects() { }
-            }
-
-            extensions.create("thing", Thing, "a")
-            assert thing.objects != null
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "can inject service using abstract getter"() {
-        buildFile << """
-            import ${Inject.name}
-
-            abstract class Thing {
-                Thing(String a) {
-                }
-
-                @Inject
-                abstract ObjectFactory getObjects()
-            }
-
-            extensions.create("thing", Thing, "a")
-            assert thing.objects != null
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "can use getter injected services from constructor"() {
-        buildFile << """
-            import ${Inject.name}
-
-            class Thing {
-                Thing(String a) {
-                    objects.property(String).set(a)
-                }
-
-                @Inject
-                ObjectFactory getObjects() { }
-            }
-
-            extensions.create("thing", Thing, "a")
-            assert thing.objects != null
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "can inject service using getter on interface"() {
-        buildFile << """
-            import ${Inject.name}
-
-            interface Thing {
-                @Inject
-                ObjectFactory getObjects()
-            }
-
-            extensions.create("thing", Thing)
-            assert thing.objects != null
-        """
-
-        expect:
-        succeeds()
-    }
-
-    @Unroll
-    def "service of type #serviceType is available for injection into project extension"() {
-        buildFile << """
-            class Thing {
-                ${serviceType} service
-
-                Thing(${serviceType} service) {
-                    this.service = service
-                }
-            }
-
-            extensions.create("thing", Thing)
-            assert thing.service != null
-        """
-
-        expect:
-        succeeds()
-
-        where:
-        serviceType << [
-            ObjectFactory,
-            ProjectLayout,
-            ProviderFactory,
-            WorkExecutor,
-            FileSystemOperations,
-            ExecOperations,
-        ].collect { it.name }
-    }
-
-    @Unroll
-    def "service of type #serviceType is available for injection into settings extension"() {
-        settingsFile << """
-            class Thing {
-                ${serviceType} service
-
-                Thing(${serviceType} service) {
-                    this.service = service
-                }
-            }
-
-            extensions.create("thing", Thing)
-            assert thing.service != null
-        """
-
-        expect:
-        succeeds()
-
-        where:
-        serviceType << [
-            ObjectFactory,
-            ProviderFactory,
-            FileSystemOperations,
-            ExecOperations,
-        ].collect { it.name }
-    }
-
-    @Unroll
-    def "service of type #serviceType is available for injection into gradle object extension"() {
-        settingsFile << """
-            class Thing {
-                ${serviceType} service
-
-                Thing(${serviceType} service) {
-                    this.service = service
-                }
-            }
-
-            gradle.extensions.create("thing", Thing)
-            assert gradle.thing.service != null
-        """
-
-        expect:
-        succeeds()
-
-        where:
-        serviceType << [
-            ObjectFactory,
-            ProviderFactory,
-            FileSystemOperations,
-            ExecOperations,
-        ].collect { it.name }
     }
 
     def "fails when construction parameters provided for interface"() {
